Add overflow menu support for plot code
